fix: reorganizes test cases so that they work propertly
