+++ server/src/repositories/users.ts
@@ -1,6 +1,4 @@
-import { PrismaClient } from "@prisma/client";
-
-const prisma = new PrismaClient();
+import prisma from "../singleton"
 
 export interface UserUpdateData {
   nickname?: string;
