+++ server/src/singleton.ts
